storage: distribute source work across all timely workers
Currently, only one worker in a `storaged` is doing the actual
ingestion/reclocking work for a source. This changes that to use all the
available timely workers, if possible. Parallelism is still bounded by
the number of "partitions".

This splits the monolithic `create_raw_source` operator into three
operators:
 - `source_reader` (parallelism = num workers): does the actual reading
   from the external system and sends along un-timestamped batches and
   summaries about source uppers
 - `remap` (parallelism = 1): takes in all the summaries about source
   uppers and periodically mints new timestamps/offsets into the remap
   shard. This sends along any update that have been written to persist.
 - `reclock` (parallelism = num workers): Takes in batches and updates
   from the `remap` operator. Uses those updates to reclock batches and
   send them along.

This introduces the new `ReclockFollower` which can ingest updates
emitted by a `ReclockOperator` and reclock batches. This can be seen as
a read-only mirror of the `ReclockOperator`. This could in theory also
directly read from persist, but I chose to actively push along updates
for now.

This introduces a new contract for downgrading capabilities to the
`SourceReader`, that contract is documented in its rustdoc.