sql: refactor load gen ast -> storage calls
Only convert from the load gen ast to load gen storage enum in a single
falliable function. This is needed by the upcoming TPCH load gen which
must be allowed to fail on invalid scale factor which has to happen when
converting from AST nodes, because afterward we don't have an error
path.

Change the generator schema to be the string version of the ast node,
future proofing it against new load generators.