storage: use all Timely workers when performing reads from persist (#14177)
* persist: correct From<Serde...> for ReaderEnrichedHollowBatch error string

* persist: add SeqNo leases to ReadHandle/ReaderEnrichedHollowBatch

* persist: persist_source generates SeqNo leases, and handles dropping them

* persist: add BatchFetcher; remove ReadHandle::fetch_batch

* persist: refactor persist_source to use mpsc, rather than mutex

* persist: add fetch submodule to read

* persist: refactor BatchFetcher to contain only necessary parts of ReadHandle, not own it

* persist: add encapsulation around SeqNo leases

* persist: improve mechanism to enforce LeasedBatch's drop semantics

* test: add SeqNo lease test

* persist: remove unnecessary Clone/serde impls on LeasedBatch

We only want `SerdeLeasedBatch` to be shipped around timely flows.

* persist: `return_lease()` should not be public

* persist: simplfy `LeasedBatch` code by inlining it

This keeps the amount of surface/code that we keep in persist at a
minimum. Plus the code that was using it in `persist_source` is not
really more complicated now.

* persist: simplify more persist code by inlining it

This also doesn't make the call-sites more complicated. Seems nice even
that you can now explicitly see what's going on, plus we don't have
these hideous `Session` parameters in those methods. \o/

* persist: fix special "droppable" logic in SerdeBatch encoding

Without this some problems in the protocol might not show up.

* persist: remove `LeasedBatch` routing machinery from persist

This removes a lot of code while the code in `persistent_source` stays
mostly at the same level of complexity. The methods we had in persist
mostly had very boilerplaty signatures and again would mostly increase
the surface/amount of things we have in persist, which I'd like to keep
small.

This change highlighted a problem in the special "droppable" encoding
logic: `SubscribeBatchReturner::return_batches()` should have panic'ed
because giving the batches to `consumed_batch_tx` calls `drop()` on
them. This was not happening before because all batches that went
through `SerdeLeasedBatch` once had their droppable field always set to
`true`. I'm solving this by passing `SerdeLeasedBatch`'es straight throw
in the "leasor" operator.

* persist: don't use unreliable in seqno test

* storage: in `persist_source`, reuse input buffers

This is the typical battern for low-level timely operators.

* persist: move making droppable LeasedBatches out of encoding

See https://github.com/MaterializeInc/materialize/pull/14177#discussion_r950562766

* persist: create new mod fetch

The fetch module under read more rightly should be a sibling module

* persist: move LeasedBatch into persist-client::fetch

* persist: move SerdeLeasedBatch to persist-client::fetch

SerdeLeasedBatch is now an important component in our story around
exchanging LeasedBatches in a way that helps users prevent leaking
SeqNos. As such, its visibility needs to change, so as to not
require making it visible through the impl::encoding module.

This commit accomplishes that goal by moving it into the fetch
module, which is made public.

* persist: refactor LeaseLifeCycle

* persist: bring seqno_leases test up to new state

* persist: fix LeasedBatch docstring

Co-authored-by: Aljoscha Krettek <aljoscha.krettek@gmail.com>