persist: return all newer versions from Consensus::compare_and_set
With incremental state, we'll need all versions in `(expected, current]`
to update our local copy of the state and try again, not just current.
It's easy to return these from our Consensus impls (and the prod one has
to make a network call here anyway), so change compare_and_set to return
exactly the set of things needed.

This quite unfortunately breaks MaelstromConsensus in a way that is
difficult to fix. This is okay because we don't need MaelstromConsensus
for correctness checking, it just gets us pretty graphs. Avoid loss of
CI coverage by replacing its use in test/persist (MemConsensus with
concurrency=2 gets us 95% of the value of true multi-node maelstrom).