catalog: Add unsigned ints to builtin tables (#14387)
Previously, Materialize did not support unsigned integers, so builtin
table columns were all signed integers. This required us represent some
data that is under the hood an unsigend int as a signed int, which also
required a lot of casts.

This commit changes all of these builtin table columns to use unsigned
ints and remove all of the casts.

This commit also updates SQL Logic Tests and Tessdrive to be able to
recognize unsigned integers.