persist: fix bug in FuelingMerge progress
We previously tracked how much total fuel a FuelingBatch had received,
calling it progress. Whenever it received new fuel, we compared progress
against self.b1.len() + self.b2.len() to see how much work remained.
However, because we (quite surprisingly) swap out the innards of a batch
when applying a compaction response, if the compaction resulted in
consolication of updates, a batch len could become smaller. This could
lead to a situation where progress was greater than the sum of the two
batch lengths. In debug, this would present as a subtraction underflow
panic. In release, it would actually underflow and result in a batch
that would not complete given any amount of fuel, which then presented
as a "Attempted to insert batch into incomplete merge" panic.

To fix, compute the amount of total progress needed once when the
FuelingBatch is constructed and subtract provided work fuel from it
until it hits zero.