Tighten compute controller public API
This commit attempts to tighten the public API of the compute controller
(`mz_compute_client::controller`). Previously, a lot of internal fields
were `pub` even though they didn't need to and probably shouldn't be.
After this change, none of the struct fields are public anymore, and all
of the public API is defined through `pub` methods.

The benefit of this refactor is that it becomes easier to reason about
how clients can and cannot affect compute controller state.