persist: temporarily disable StateDiff::validate_roundtrip
This is an internal validation that is only run in debug mode. It
attempts to assert that, given the state, new_state, and diff generated
in the core apply_unbatched_cmd loop, that applying diff to state
results in something equal to new_state. Sadly, when applying the diff
for a compaction apply_merge_res, we currently fall back to the "slow
path", which involved deconstructing a spine into HollowBatches,
applying the diff to the raw HollowBatches, and then creating a new
Spine from scratch. This isn't guaranteed to result in the same
structure as when we started because of things like the optimization for
merging empty batches.

For now, we'll disable the validation. I think the fix has to be fixing
the compaction fast path in apply_diff (concretely: shaking out the
issues that caused it from being used in the original version of inc
state). After we do that, we'll have more guarantees about the structure
of the spine after applying a diff.

Touches #14490