persist: don't use proto to serialize/deserialize SerdeLeasedBatch
This was a bit of over-engineering on my part. Timely clusters are, and
always will be, shared fate and thus run the same version of code.
Persist then doesn't need to worry about backward/forward compatibility
in its ExchangeData impl and so this protobuf dance is wasted work and
cycles. Instead, derive Serialize and Deserialize as normal.

While I'm in here, tighten up the fields of LeasedBatch to be more part
specific in preparation for the next commit, which renames them (but
pull this out separately so the next commit is a pure rename).

Also rename `until` in the LeasedBatchMetadata to `lower`. `until` has a
specific meaning in differential dataflow that I wasn't aware of when
this name was picked. The usual definition of `lower` matches the
semantics of the field, so use that.