Introduce `ActiveComputeController`
Previously, the state of a (per-instance) compute controller was kept in
a `ComputeControllerState` struct. Two wrapper structs,
`ComputeController` and `ComputeControllerMut` existed to bundle this
state with a `&mut StorageController`, and provide immutable and mutable
access to the controller state, respectively. The environment-global
`Controller` kept a collection of `ComputeControllerState`s, and
converted those into `ComputeController`s/`ComputeControllerMut`s as
necessary.

This commit changes this approach by replacing the above structs with
two new ones: `ComputeController` and `ActiveComputeController`.
`ComputeController` is what was previously the `ComputeControllerState`
(plus the instance ID), whereas `ActiveComputeController` is what was
previously `ComputeControllerMut`. The old `ComputeController` has been
replaced by `&ComputeController`, no separate struct is required.

The main motivation for this change is that it simplifies the
implementation of durable compute controller commands. Some parts of the
controller state (e.g. `stashed_response`) should not be made durable,
and the new design makes factoring out the durable part of the state
more straightforward. Aside from that, the new approach is also
subjectively more intuitive, and it is consistent with the approach we
already use in `computed` (see `ComputeState` and `ActiveComputeState`).