persist: special-case a shard with empty upper and empty since
If the upper and shard-global since are both the empty antichain, then
no further writes can ever commit and no further reads can be served.
Optimize the shard into a final unchanging tombstone state by replacing
the spine with a single empty batch and clearing the reader and writer
registrations.

Also included are a couple tweaks that I felt helped make this easier to
reason about:
- Eliminate the clone_reader cmd. It was exactly the same impl as
  register_reader. Instead, we replace it with an assert in
  ReadHandle::clone that the since of the new registration is <= the
  since of the handle being cloned.
- Make apply_unbatched_idempotent_cmd return optional maintenance for
  no-op cmd application on tombstone shards.