storage: in upsert operator, fix hashing of previous and current updates
It's very important to hash the right thing. We have nested `Option` and
`Result` here. And, for example, `Some(key).hashed()` is not the same as
`key.hashed()`.

This change makes sure that we hash the same thing for both pervious
updates and new updates.

Also: this problem was only showing up when trying to use upsert-style
sources with multiple storaged workers.