compute controller: respect upper lower bounds
This commit changes the compute controller to base the calculation of
the implied read capability of a collection (what determines how far the
inputs to the collection's dataflow can be compacted) on the lower bound
of all reported upper frontiers, rather than the upper bound. This
ensures that compaction is held back until the slowest replica is done
reading at the respective times, and thereby prevents crashes due to
attempted reads at already compacted times.