replicated: report lower upper bounds for tails
This commit makes the `ActiveReplication` client report the lower bounds
of upper frontiers reported by all replicas for in-progress tails, in
addition to the upper bounds that were already reported before.

This ensures that compaction of tail sources is held back far enough for
the slowest replica to still be able to successfully read at the times
it needs to.