compute-client: fix multi-process frontier tracking
The `Partitioned` compute client is not guaranteed to see all compute
commands in a multi-process replica. Thus, we cannot rely on observed
commands to initialize frontier tracking.

In this commit, frontier tracking is initialized based on received
`FrontierUppers` responses instead.

Frontier tracking state is cleaned up when all shards have reported an
empty frontier for a collection. For this to always work, we need to
ensure that the compute workers emit final `FrontierUppers` responses in
response to `AllowCompaction` commands that allow collections to be
compacted to the empty frontier.