adapter: add metrics for query counts and active sessions/subscribes
This commit provides most of the metrics requested in
MaterializeInc/cloud#4146. It ends up being much easier to track these
metrics in the coordinator rather than the pgwire crate, so that we have
visibility into queries issued both via HTTP and pgwire.

Note that the active subscribes metric is currently broken and never
decrements, due to a pre-existing bug that causes the coordinator to
never hear about dropped subscribes.